requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
datasets>=2.14.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Utilities
tqdm>=4.65.0
numpy>=1.21.0
matplotlib>=3.5.0
pyyaml>=6.0
requests>=2.28.0

# Optional for development
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.990
